9: part 3 - increase score to u32, pickup score items

+++ common/variables.h
@@ -50,7 +50,7 @@ struct game_state
   u8 tick;
   u8 dave_tick;
   u8 current_level;
-  u8 score;
+  u32 score;
 	u8 lives;
   u8 view_x;
   u8 view_y;
